+++ doorche-admin/src/main/webapp/WEB-INF/views/sm/role/list.jsp
@@ -15,6 +15,7 @@
     </ul>
     <div class="tab-content">
         <div class="tab-pane active">
+            <sys:message content="${message}" type="${messageType}"/>
             <div class="box">
                 <div class="box-header with-border">
                     <h3 class="box-title">查询条件</h3>
@@ -83,7 +84,7 @@
                                         <a href="${ctx}/sm/role/edit?id=${role.id}">修改</a>
                                     </security:hasPermission>
                                     <security:hasPermission name="sys:role:delete">
-                                        <a href="${ctx}/sm/role/delete?id=${role.id}">删除</a>
+                                        <a href="${ctx}/sm/role/delete?id=${role.id}" onclick="return $.windowBox.confirm('确认删除该角色吗？',this.href)">删除</a>
                                     </security:hasPermission>
                                 </td>
                             </tr>
